Domain Model review
