+++ samples/Dom.Mediator.Samples.MinimalApi/Features/CreateTask/CreateTaskCommand.cs
@@ -2,5 +2,5 @@ using Dom.Mediator.Abstractions;
 
 public record CreateTaskCommand(
     string Title, 
-    string? Description, 
-    DateTime? DueDate) : ICommand<string>;
+    string Description, 
+    DateTime DueDate) : ICommand<string>;
