+++ samples/Dom.Mediator.Samples.MinimalApi/Features/GetAllTasks/GetAllTasksHandler.cs
@@ -1,5 +1,6 @@
 using Dom.Mediator;
 using Dom.Mediator.Abstractions;
+using Dom.Mediator.Samples.MinimalApi.Features;
 
 public class GetAllTasksHandler : IQueryHandler<GetAllTasksQuery, List<TaskItem>>
 {
