+++ samples/Dom.Mediator.Samples.MinimalApi/Features/GetAllTasks/GetAllTasksQuery.cs
@@ -1,3 +1,4 @@
 using Dom.Mediator.Abstractions;
+using Dom.Mediator.Samples.MinimalApi.Features;
 
 public record GetAllTasksQuery() : IQuery<List<TaskItem>>;
