+++ samples/Dom.Mediator.Samples.MinimalApi/Features/UpdateTask/UpdateTaskCommand.cs
@@ -1,4 +1,5 @@
 using Dom.Mediator.Abstractions;
+using Dom.Mediator.Samples.MinimalApi.Features;
 
 public record UpdateTaskCommand(
     string Id,
